Consolidate unreleased Execution API versions under 2026-04-06
