+++ airflow-core/src/airflow/api_fastapi/execution_api/versions/__init__.py
@@ -28,31 +28,30 @@ from airflow.api_fastapi.execution_api.versions.v2025_08_10 import (
 from airflow.api_fastapi.execution_api.versions.v2025_09_23 import AddDagVersionIdField
 from airflow.api_fastapi.execution_api.versions.v2025_10_27 import MakeDagRunConfNullable
 from airflow.api_fastapi.execution_api.versions.v2025_11_05 import AddTriggeringUserNameField
-from airflow.api_fastapi.execution_api.versions.v2025_11_07 import AddPartitionKeyField
-from airflow.api_fastapi.execution_api.versions.v2025_12_08 import (
+from airflow.api_fastapi.execution_api.versions.v2026_04_06 import (
+    AddDagEndpoint,
     AddDagRunDetailEndpoint,
-    MovePreviousRunEndpoint,
-)
-from airflow.api_fastapi.execution_api.versions.v2026_03_31 import (
     AddNoteField,
+    AddPartitionKeyField,
     MakeDagRunStartDateNullable,
     ModifyDeferredTaskKwargsToJsonValue,
+    MovePreviousRunEndpoint,
     RemoveUpstreamMapIndexesField,
 )
-from airflow.api_fastapi.execution_api.versions.v2026_04_13 import AddDagEndpoint
 
 bundle = VersionBundle(
     HeadVersion(),
-    Version("2026-04-13", AddDagEndpoint),
     Version(
-        "2026-03-31",
+        "2026-04-06",
+        AddPartitionKeyField,
+        MovePreviousRunEndpoint,
+        AddDagRunDetailEndpoint,
         MakeDagRunStartDateNullable,
         ModifyDeferredTaskKwargsToJsonValue,
         RemoveUpstreamMapIndexesField,
         AddNoteField,
+        AddDagEndpoint,
     ),
-    Version("2025-12-08", MovePreviousRunEndpoint, AddDagRunDetailEndpoint),
-    Version("2025-11-07", AddPartitionKeyField),
     Version("2025-11-05", AddTriggeringUserNameField),
     Version("2025-10-27", MakeDagRunConfNullable),
     Version("2025-09-23", AddDagVersionIdField),
